+++ src/main/java/domain/lotto/Ball.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lotto;
 
 public class Ball {
