+++ src/main/java/domain/lotto/Lottos.java
@@ -1,9 +1,9 @@
-package domain;
+package domain.lotto;
 
-import static domain.Lotto.LOTTO_PRICE;
-import static domain.Lotto.MAX_LOTTO_NUMBER;
-import static domain.Lotto.MAX_LOTTO_SIZE;
-import static domain.Lotto.MIN_LOTTO_NUMBER;
+import static domain.lotto.Lotto.LOTTO_PRICE;
+import static domain.lotto.Lotto.MAX_LOTTO_NUMBER;
+import static domain.lotto.Lotto.MAX_LOTTO_SIZE;
+import static domain.lotto.Lotto.MIN_LOTTO_NUMBER;
 
 import java.util.ArrayList;
 import java.util.Comparator;
