+++ src/main/java/domain/winning/WinningStatistics.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.winning;
 
 public enum WinningStatistics {
