+++ src/main/java/domain/winning/Yield.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.winning;
 
 import java.util.List;
