+++ src/test/java/domain/LottoTest.java
@@ -3,6 +3,9 @@ package domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import domain.lotto.Lotto;
+import domain.winning.Matcher;
+import domain.winning.WinningLotto;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
