+++ src/test/java/domain/LottosTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import domain.lotto.Lottos;
 import org.junit.jupiter.api.Test;
 
 class LottosTest {
