+++ src/test/java/view/InputConverterTest.java
@@ -26,7 +26,7 @@ class InputConverterTest {
         assertThat(winningNumbers).isEqualTo(List.of(1, 2, 3, 4, 5, 6));
     }
 
-    @ValueSource(strings = {"a", "-1", "1,,2", "1,a", "1,1,2,3,4,5"})
+    @ValueSource(strings = {"a", "-1", "1,,2", "1,a"})
     @ParameterizedTest
     void 당첨_번호_입력_값이_올바르지_않다면_예외를_던진다(String input) {
